+++ CHANGELOG.md
@@ -10,7 +10,11 @@
   + `FirstClassReflection`, which must be enabled to define a `%reflection`
     function
 
+## Library Updates
 
++ Terminating programs has been improved with more appropriate
+  functions (`exitWith`, `exitFailure`, and `exitSuccess`) and a data
+  structure (`ExitCode`) to capture a program's return code.
 
 # New in 0.99: